Use bicubic and bilinear interpolation for height calculations

+++ osm_db/src/area_db.rs
@@ -660,7 +660,7 @@ impl AreaDatabase {
         }
     }
 
-    pub fn elevation_at_coords(&self, lat: f64, lon: f64) -> Option<i16> {
+    pub fn elevation_at_coords(&self, lat: f64, lon: f64) -> Option<f64> {
         self
             .elevation_map
             .as_ref()
