+++ osm_db_py/src/area_db.rs
@@ -128,7 +128,7 @@ impl PyAreaDatabase {
         }
     }
 
-    pub fn elevation_at_coords(&self, lat: f64, lon: f64) -> Option<i16> {
+    pub fn elevation_at_coords(&self, lat: f64, lon: f64) -> Option<f64> {
         self.inner.lock().unwrap().elevation_at_coords(lat, lon)
     }
 }
